+++ LightBulb/Function/EuclideanDistance.cpp
@@ -4,7 +4,7 @@
 #include "Neuron\RBFThreshold.hpp"
 #include <stdexcept>
 
-float EuclideanDistance::execute(std::list<Edge*> &input, AbstractThreshold* threshold, float additionalInput, std::map<AbstractNeuron*, float>* neuronOutputCache)
+double EuclideanDistance::execute(std::list<Edge*> &input, AbstractThreshold* threshold, double additionalInput, std::map<AbstractNeuron*, double>* neuronOutputCache)
 {
 	RBFThreshold* rbfThreshold = dynamic_cast<RBFThreshold*>(threshold);
 
@@ -12,10 +12,10 @@ float EuclideanDistance::execute(std::list<Edge*> &input, AbstractThreshold* thr
 	if (!rbfThreshold)
 		std::invalid_argument("The threshold has to be a RBFThreshold");
 
-	float euclideanDistance = 0.0;
+	double euclideanDistance = 0.0;
 
 	// Calculate the square of the distance between the output and the center
-	std::vector<float>::iterator centerCoordinate = (*rbfThreshold->getCenterVector()).begin();
+	std::vector<double>::iterator centerCoordinate = (*rbfThreshold->getCenterVector()).begin();
 	for (std::list<Edge*>::iterator edge = input.begin(); edge != input.end(); edge++, centerCoordinate++)
 	{
 		euclideanDistance += pow((neuronOutputCache == NULL ? (*edge)->getPrevNeuron()->getActivation() : (*neuronOutputCache)[(*edge)->getPrevNeuron()]) - (*centerCoordinate), 2);
